remove parent Events, rename EventsTable to Events
